// server/enums/abilityToWorkValue.ts
enum AbilityToWorkValue {
  EDUCATION_ENROLLMENT = 'EDUCATION_ENROLLMENT',
  FAMILY_ISSUES = 'FAMILY_ISSUES',
  FULL_TIME_CARER = 'FULL_TIME_CARER',
  MENTAL_HEALTH_ISSUES = 'MENTAL_HEALTH_ISSUES',
  PHYSICAL_HEALTH_ISSUES = 'PHYSICAL_HEALTH_ISSUES',
  DEPENDENCY_ISSUES = 'DEPENDENCY_ISSUES',
  NONE = 'NONE',
}

export default AbilityToWorkValue